refactor: 템플릿을 찾는 로직을 mapper에서 하도록 변경
