+++ backend/src/main/java/reviewme/template/service/TemplateService.java
@@ -6,29 +6,24 @@ import org.springframework.transaction.annotation.Transactional;
 import reviewme.review.domain.exception.ReviewGroupNotFoundByReviewRequestCodeException;
 import reviewme.reviewgroup.domain.ReviewGroup;
 import reviewme.reviewgroup.repository.ReviewGroupRepository;
-import reviewme.template.domain.Template;
-import reviewme.template.domain.exception.TemplateNotFoundByReviewGroupException;
-import reviewme.template.repository.TemplateRepository;
 import reviewme.template.service.dto.response.TemplateResponse;
+import reviewme.template.service.module.TemplateMapper;
 
 @Service
 @RequiredArgsConstructor
 public class TemplateService {
 
     private final ReviewGroupRepository reviewGroupRepository;
-    private final TemplateRepository templateRepository;
     private final TemplateMapper templateMapper;
 
     @Transactional(readOnly = true)
     public TemplateResponse generateReviewForm(String reviewRequestCode) {
-        ReviewGroup reviewGroup = reviewGroupRepository.findByReviewRequestCode(reviewRequestCode)
+        ReviewGroup reviewGroup = findReviewGroupByRequestCodeOrThrow(reviewRequestCode);
+        return templateMapper.mapToTemplateResponse(reviewGroup);
+    }
+
+    private ReviewGroup findReviewGroupByRequestCodeOrThrow(String reviewRequestCode) {
+        return reviewGroupRepository.findByReviewRequestCode(reviewRequestCode)
                 .orElseThrow(() -> new ReviewGroupNotFoundByReviewRequestCodeException(reviewRequestCode));
-
-        Template template = templateRepository.findById(reviewGroup.getTemplateId())
-                .orElseThrow(() -> new TemplateNotFoundByReviewGroupException(
-                        reviewGroup.getId(), reviewGroup.getTemplateId()
-                ));
-
-        return templateMapper.mapToTemplateResponse(reviewGroup, template);
     }
 }
